backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
anthropic==0.7.1
requests==2.31.0
sqlalchemy==2.0.23
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0
structlog==23.2.0
python-json-logger==2.0.7
pyyaml==6.0
aiosqlite==0.19.0
jsonschema==4.20.0
